Remove dependencies on apt.canaltp.local repo
This is achieved by building the following dependencies from source:
 * protobuf v.3.3.1
 * osm-protobuf v.1.3.3

The version being used choosen are based on the version available on Dev
and Jenkins.

The main drawback to this approach is that we don't have way to
replicate any futher changes made to debian8 package onto our production
environement. So we should not update any of those libraries on debian8
unless we make sure they could be deployed by the DevOps.